Make some configs inner classes
