+++ SoftwareBot/src/main/java/frc/robot/utility/config/GyroConfig.java
@@ -3,12 +3,14 @@ package frc.robot.utility.config;
 import com.fasterxml.jackson.annotation.JsonSubTypes;
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 import frc.robot.subsystems.drivetrain.GyroIO;
+import frc.robot.subsystems.drivetrain.GyroIOPigeon2;
+import frc.robot.subsystems.drivetrain.GyroIOSim;
 import lombok.Data;
 
 @Data
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME)
-@JsonSubTypes({@JsonSubTypes.Type(value = Pigeon2Config.class, name = "Pigeon2"),
-@JsonSubTypes.Type(value = SimGyroConfig.class, name = "Sim")})
+@JsonSubTypes({@JsonSubTypes.Type(value = GyroIOPigeon2.Config.class, name = "Pigeon2"),
+@JsonSubTypes.Type(value = GyroIOSim.Config.class, name = "Sim")})
 public abstract class GyroConfig {
 
     public abstract GyroIO createIO();
